ref(cardinality): Moves limit state into its own module
