+++ relay-cardinality/src/redis/limiter.rs
@@ -4,16 +4,13 @@ use relay_redis::{Connection, RedisPool};
 use relay_statsd::metric;
 
 use crate::{
-    limiter::{Entry, EntryId, Limiter, Rejections, Scoping},
+    limiter::{Entry, Limiter, Rejections, Scoping},
     redis::{
         cache::{Cache, CacheOutcome},
-        quota::QuotaScoping,
         script::{CardinalityScript, Status},
+        state::{LimitState, RedisEntry},
     },
-    statsd::{
-        CardinalityLimiterCounters, CardinalityLimiterHistograms, CardinalityLimiterSets,
-        CardinalityLimiterTimers,
-    },
+    statsd::{CardinalityLimiterHistograms, CardinalityLimiterTimers},
     CardinalityLimit, Result,
 };
 use relay_common::time::UnixTimestamp;
@@ -175,136 +172,6 @@ impl Limiter for RedisSetLimiter {
     }
 }
 
-/// Internal state combining relevant entries for the respective quota.
-struct LimitState<'a> {
-    /// Entries which are relevant for the quota.
-    pub entries: Vec<RedisEntry>,
-    /// Scoping of the quota.
-    pub scope: QuotaScoping,
-    /// The limit of the quota.
-    pub limit: u64,
-
-    /// The original cardinality limit.
-    cardinality_limit: &'a CardinalityLimit,
-    /// The original/full scoping.
-    scoping: Scoping,
-    /// Amount of cache hits.
-    cache_hits: i64,
-    /// Amount of cache misses.
-    cache_misses: i64,
-    /// Amount of accepts,
-    accepts: i64,
-    /// Amount of rejections.
-    rejections: i64,
-}
-
-impl<'a> LimitState<'a> {
-    pub fn new(scoping: Scoping, cardinality_limit: &'a CardinalityLimit) -> Option<Self> {
-        Some(Self {
-            entries: Vec::new(),
-            scope: QuotaScoping::new(scoping, cardinality_limit)?,
-            limit: cardinality_limit.limit,
-            cardinality_limit,
-            scoping,
-            cache_hits: 0,
-            cache_misses: 0,
-            accepts: 0,
-            rejections: 0,
-        })
-    }
-
-    pub fn from_limits(scoping: Scoping, limits: &'a [CardinalityLimit]) -> Vec<Self> {
-        limits
-            .iter()
-            .filter_map(|limit| LimitState::new(scoping, limit))
-            .collect::<Vec<_>>()
-    }
-
-    pub fn id(&self) -> &'a str {
-        &self.cardinality_limit.id
-    }
-
-    pub fn cardinality_limit(&self) -> &'a CardinalityLimit {
-        self.cardinality_limit
-    }
-
-    pub fn take_entries(&mut self) -> Vec<RedisEntry> {
-        std::mem::take(&mut self.entries)
-    }
-
-    pub fn cache_hit(&mut self) {
-        self.cache_hits += 1;
-    }
-
-    pub fn cache_miss(&mut self) {
-        self.cache_misses += 1;
-    }
-
-    pub fn accepted(&mut self) {
-        self.accepts += 1;
-    }
-
-    pub fn rejected(&mut self) {
-        self.rejections += 1;
-    }
-}
-
-impl<'a> Drop for LimitState<'a> {
-    fn drop(&mut self) {
-        let passive = if self.cardinality_limit.passive {
-            "true"
-        } else {
-            "false"
-        };
-
-        metric!(
-            counter(CardinalityLimiterCounters::RedisCacheHit) += self.cache_hits,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-        metric!(
-            counter(CardinalityLimiterCounters::RedisCacheMiss) += self.cache_misses,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-        metric!(
-            counter(CardinalityLimiterCounters::Accepted) += self.accepts,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-        metric!(
-            counter(CardinalityLimiterCounters::Rejected) += self.rejections,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-        );
-
-        let organization_id = self.scoping.organization_id as i64;
-        let status = if self.rejections > 0 { "limited" } else { "ok" };
-        metric!(
-            set(CardinalityLimiterSets::Organizations) = organization_id,
-            id = &self.cardinality_limit.id,
-            passive = passive,
-            status = status,
-        )
-    }
-}
-
-/// Entry used by the Redis limiter.
-#[derive(Clone, Copy, Debug)]
-struct RedisEntry {
-    /// The correlating entry id.
-    pub id: EntryId,
-    /// The entry hash.
-    pub hash: u32,
-}
-
-impl RedisEntry {
-    /// Creates a new Redis entry.
-    fn new(id: EntryId, hash: u32) -> Self {
-        Self { id, hash }
-    }
-}
-
 struct CheckedLimits {
     entries: Vec<RedisEntry>,
     statuses: Vec<Status>,
@@ -333,6 +200,7 @@ mod tests {
     use relay_base_schema::project::ProjectId;
     use relay_redis::{redis, RedisConfigOptions};
 
+    use crate::limiter::EntryId;
     use crate::redis::{KEY_PREFIX, KEY_VERSION};
     use crate::{CardinalityScope, SlidingWindow};
